Don't use CentralSystem from ChargePoint
